Remove simplest unit of work base implementation
