+++ mb/unit_of_work/__init__.py
@@ -1,2 +1,1 @@
-from .base import UnitOfWork  # noqa
-from .nested import NestedUnitOfWork  # noqa
+from .uow import UnitOfWork  # noqa
